+++ config/packages/slepc4py.py
@@ -15,31 +15,44 @@ class Slepc4py(package.Package):
   def __init__(self,argdb,log):
     package.Package.__init__(self,argdb,log)
     self.packagename     = 'slepc4py'
-    self.downloadable    = True
-    self.version         = 'master'
-    self.url             = 'https://gitlab.com/slepc/slepc4py.git'
-    self.dirname         = 'slepc4py-'+self.version
-    self.builtafterslepc = True
+    self.installable     = True
     self.ProcessArgs(argdb)
 
+  def ProcessArgs(self,argdb,petscpackages=''):
+    url,flag,found = argdb.PopUrl('download-slepc4py')
+    if found:
+      self.log.Exit('--download-slepc4py has been renamed to --with-slepc4py\nUse -h for help')
+    value,found = argdb.PopBool('with-'+self.packagename)
+    if found:
+      self.requested = value
 
-  def DownloadOnly(self,slepcconf,slepcvars,slepcrules,slepc,petsc,archdir,prefixdir):
-    externdir = slepc.CreateDir(archdir,'externalpackages')
-    destdir   = os.path.join(prefixdir,'lib')
+  def ShowHelp(self):
+    wd = package.Package.wd
+    print('  --with-slepc4py=<bool>'.ljust(wd)+': Build Python bindings (default: no)')
 
-    # Check if source is already available
-    builddir = os.path.join(externdir,self.dirname)
-    if os.path.exists(builddir):
-      self.log.write('Using '+builddir)
-    else: # clone slepc4py repo
-      url = self.packageurl
-      if url=='':
-        url = self.url
-      self.log.write('Cloning '+url+' to '+builddir)
-      try:
-        (result,output) = self.RunCommand('cd '+externdir+'&& git clone -b '+self.version+' '+url+' '+self.dirname)
-      except RuntimeError as e:
-        self.log.Exit('Cannot clone '+url+': '+str(e))
+  def ShowInfo(self):
+    if self.havepackage:
+      self.log.Println('Python bindings (slepc4py) will be built after SLEPc')
+
+  def Process(self,slepcconf,slepcvars,slepcrules,slepc,petsc,archdir=''):
+    if not self.requested:
+      self.SkipInstall(slepcrules)
+      return
+    self.log.NewSection('Processing slepc4py...')
+
+    # Check petsc4py module
+    try:
+      from petsc4py import PETSc
+    except ImportError:
+      self.log.Exit('Cannot import petsc4py, make sure your PYTHONPATH is set correctly')
+    # Check for cython
+    try:
+      import Cython
+    except ImportError:
+      self.log.Exit('--with-slepc4py requires that cython is installed on your system')
+
+    builddir = os.path.join(slepc.dir,'src','binding','slepc4py')
+    destdir  = os.path.join(slepc.prefixdir,'lib')
 
     # add makefile rules
     envvars = 'PETSC_ARCH="" SLEPC_DIR=${SLEPC_INSTALLDIR}' if slepc.isinstall else ''