+++ CHANGELOG.md
@@ -4,10 +4,12 @@
 
 ### Added
 
+- Import SLEPc4py sources into SLEPc source tree.
 - New tolerance to check a definite inner product, see `BVSetDefiniteTolerance()`.
 
 ### Changed
 
+- Configure option `--download-slepc4py` renamed to `--with-slepc4py`.
 - `ST`: now the coefficient/preconditioner matrix has the same prefix as the `KSP` object,
   so setting MUMPS options in the command-line requires the appropriate prefix, e.g.,
   `-st_mat_mumps_icntl_13 1`.